IMPLEMENT - the event notification new implement.
Change naming convention: publisher --> subject, subscriber --> observer

the event module do not store statically allocate memory for the subject
and observer. The application module, must provide those structure to
the event module. There is an disadvantage in this implementation,
the structure of the observer and subject are exposed to the application
user.